requirements.txt
camoufox[geoip]
playwright
beautifulsoup4
markdownify
pyyaml
httpx
